Fix Visual Studio warnings with FPL libraries.
FPL libraries and Zooshi now compile without warnings
on Visual Studio 2010 and 2015. Note that Zooshi does not
link on VS2010 because the Firebase C++ SDK only comes with
libs for VS2015 at the moment (see b/35421579).
PiperOrigin-RevId: 147723846

+++ src/invites.cpp
@@ -50,8 +50,8 @@ bool UpdateSentInviteStatus(bool* did_send, bool* first_sent) {
         *first_sent = invite_sent_count == 0;
       }
       fplbase::SavePreference(
-          kInviteSentKey,
-          invite_sent_count + future.Result()->invitation_ids.size());
+          kInviteSentKey, static_cast<int32_t>(
+              invite_sent_count + future.Result()->invitation_ids.size()));
     }
     return true;
   } else {
